requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
pkg-resources==0.0.0
requests==2.25.1
ruamel.yaml==0.16.12
typer==0.3.2
urllib3==1.26.2
